Update Delete stored procedures

+++ src/CombatAnalysis.BL/Services/DamageDoneService.cs
@@ -43,6 +43,15 @@ namespace CombatAnalysis.BL.Services
             return DeleteInternalAsync(item);
         }
 
+        async Task<int> IService<DamageDoneDto>.DeleteByProcedureAsync(int combatPlayerId)
+        {
+            var paramNames = new string[] { nameof(combatPlayerId) };
+            var paramValues = new object[] { combatPlayerId };
+
+            var response = await _repository.DeleteByProcedureAsync(DbProcedureHelper.DeleteDamageDone, paramNames, paramValues);
+            return response;
+        }
+
         async Task<IEnumerable<DamageDoneDto>> IService<DamageDoneDto>.GetAllAsync()
         {
             var allData = await _repository.GetAllAsync();
@@ -90,11 +99,11 @@ namespace CombatAnalysis.BL.Services
 
         private async Task<int> DeleteInternalAsync(DamageDoneDto item)
         {
-            //var allData = await _repository.GetAllAsync();
-            //if (!allData.Any())
-            //{
-            //    throw new NotFoundException($"Collection entity {nameof(DamageDoneDto)} not found", nameof(allData));
-            //}
+            var allData = await _repository.GetAllAsync();
+            if (!allData.Any())
+            {
+                throw new NotFoundException($"Collection entity {nameof(DamageDoneDto)} not found", nameof(allData));
+            }
 
             var numberEntries = await _repository.DeleteAsync(_mapper.Map<DamageDone>(item));
             return numberEntries;
@@ -110,11 +119,6 @@ namespace CombatAnalysis.BL.Services
 
             var numberEntries = await _repository.UpdateAsync(_mapper.Map<DamageDone>(item));
             return numberEntries;
-        }
-
-        public Task<int> DeleteByProcedureAsync(int combatPlayerId)
-        {
-            throw new NotImplementedException();
         }
     }
 }
